+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Trigger a default search when the search form is reset.
+
 ## v4.10.0 (2023-10-30)
 
 - Force a default search when initially viewing the application interface.